Time: 64 ms (76.5%), Space: 148.6 MB (64.55%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.cpp
@@ -40,7 +40,7 @@ public:
 
         if(left_max.size() == right_min.size())
         {
-            return (double)((left_max.top()+right_min.top())/2.0);
+            return((left_max.top()+right_min.top())/2.0);
         }
         else
             return left_max.top();
